backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.27.0
python-multipart==0.0.9
pydantic==2.7.3
pydantic-settings==2.2.1
python-docx==1.1.0
Pillow==10.3.0
